chore(ansible): move _SECRET_TO_ANSIBLE_VAR to shared location (#3778)
